chore: move page pool out of autoload
